+++ CHANGELOG.md
@@ -7,9 +7,15 @@ and this project adheres to [0ver](https://0ver.org) (more or less).
 
 ## [Unreleased]
 
+### Added
+
+- New `updown_webhook` resource to manage webhooks configuration
+
 ### Changed
 
 - Switched libs to github.com/hashicorp/terraform-plugin-sdk instead of formerly github.com/hashicorp/terraform
+- Upgraded to go 1.17
+- Bumped dependencies to their latest versions
 
 ## [0.2.1] - 2020-09-11